% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_extract_from_fit.R
\name{add_info_fit_desc}
\alias{add_info_fit_desc}
\title{Fit description customization}
\usage{
add_info_fit_desc(
  fits,
  exp = TRUE,
  digits = 2,
  boundknots = FALSE,
  custom_model_label = TRUE
)
}
\arguments{
\item{fits}{{\code{list}}\cr flexsurvreg, flexsurvpline fit object}

\item{exp}{{\code{logical}}\cr Boolean to indicate if exponential transformation
should be used on knots location values}

\item{digits}{{\code{integer}}\cr Indicate the number of decimal places to use when
rounding knots location values see \code{\link[base:Round]{base::round()}}}

\item{custom_model_label}{{\code{logical}}\cr Boolean to indicate if spline knots
locations should be included in the model label}
}
\value{
Character vector with fit name with knot location added
}
\description{
Fit description customization
}
\section{Specification}{

\if{latex}{
Add knots location to fit name. 

Logical Specifications:
  \enumerate{
  \item Bulletproof check fit object list using \code{\link{bullet_proof}}
  \item Loop through the fits, extract knot location information and update
      fit name. If fit$knots exists:
       \enumerate{
            \item Extract knot(s)' location information from fit$knots
            \item keep only internal knot(s) - remove first and last 
            \item round knot location value with round(...,digits=digits) function
            \item concatenate fit name and knot location information using 
                glue::glue_collapse(...,sep=', ')
        }
  \item Output updated name vector
  }
}
\if{html}{
The contents of this section are shown in PDF user manual only.
}
}

\keyword{internal}
