% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_extract_from_fit.R
\name{add_log_rest_row}
\alias{add_log_rest_row}
\title{Row header customization}
\usage{
add_log_rest_row(fit)
}
\arguments{
\item{fit}{{\code{flexsurvreg}}\cr flexsurvreg fit object}
}
\value{
Character vector with updated row names
}
\description{
Row header customization
}
\section{Specification}{

\if{latex}{
Add log(..) text to row name if value are on log scale. 

Logical Specifications:
  \enumerate{
  \item Compare res and res.t value object. If fit$res and fit$res.t exist:
       \enumerate{
            \item Extract row names
            \item Loop through rows of res.t/res to compare values in first
                  column. If res value is equal to exponential of res.t value
                  then update rowname to add log using glue::glue("log({rowname}))
        }
  \item Output updated row name vector
  }
}
\if{html}{
The contents of this section are shown in PDF user manual only.
}
}

\keyword{internal}
