% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{adsl}
\alias{adsl}
\title{Dummy Subject Level-Dataset}
\format{
A data frame with 400 rows and 55 variables
}
\usage{
data(adsl)
}
\description{
See https://github.com/insightsengineering/random.cdisc.data/blob/main/R/radsl.R
for more details.
}
\keyword{datasets}
