% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{adtte}
\alias{adtte}
\title{Dummy Time-to-Event Dataset}
\format{
A data frame with 2000 rows and 67 variables
}
\usage{
data(adtte)
}
\description{
See https://github.com/insightsengineering/random.cdisc.data/blob/main/R/radtte.R
for more details.
}
\keyword{datasets}
