% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_extract_from_fit.R
\name{extract_est}
\alias{extract_est}
\title{Extract Survival Time and Estimates}
\usage{
extract_est(
  fits = list(),
  t = NULL,
  ci = FALSE,
  se = FALSE,
  B = 1000,
  cl = 0.95,
  na.action = na.pass,
  exp = TRUE,
  digits = 2,
  seed = 1234,
  custom_model_label = TRUE
)
}
\arguments{
\item{fits}{{\code{list}}\cr flexsurvreg, flexsurvpline fit object}

\item{t}{{\code{numeric}}\cr numeric vector for time interval to extract\cr
see t parameter in \code{\link[flexsurv:summary.flexsurvreg]{flexsurv::summary.flexsurvreg()}}}

\item{ci}{{\code{logical}}\cr boolean to include confidence interval estimates\cr
see ci parameter in \code{\link[flexsurv:summary.flexsurvreg]{flexsurv::summary.flexsurvreg()}}}

\item{se}{{\code{logical}}\cr boolean to include standard error estimates\cr
see se parameter in \code{\link[flexsurv:summary.flexsurvreg]{flexsurv::summary.flexsurvreg()}}}

\item{B}{{\code{numeric}}\cr Number of simulations from the normal asymptotic\cr
distribution see B parameter in \code{\link[flexsurv:summary.flexsurvreg]{flexsurv::summary.flexsurvreg()}}}

\item{cl}{{\code{numeric}}\cr Width of symmetric confidence intervals, relative to 1\cr
see cl parameter in \code{\link[flexsurv:summary.flexsurvreg]{flexsurv::summary.flexsurvreg()}}}

\item{na.action}{Function determining what should be done with missing values\cr
see na.action parameter in \code{\link[flexsurv:summary.flexsurvreg]{flexsurv::summary.flexsurvreg()}}}

\item{exp}{{\code{logical}}\cr Boolean to indicate if exponential transformation
should be used on knots location values}

\item{digits}{{\code{integer}}\cr Indicate the number of decimal places to use when
rounding knots location values see \code{\link[base:Round]{base::round()}}}

\item{seed}{seed to use to allow reproducibility of estimates}

\item{custom_model_label}{{\code{logical}}\cr Boolean to indicate if spline knots
locations should be included in the model label}
}
\value{
list with survival and hazard estimates information
}
\description{
Extract Survival Estimates information from fit objects.
}
\section{Specification}{

\if{latex}{
Extract Survival Estimates information from fit objects. 

Logical Specifications:
  \enumerate{
  \item Bulletproof check fit object list using \code{\link{bullet_proof}}
  \item Update name of list elements using \code{\link{add_info_fit_desc}} with
  parameter exp and digits
  \item Loop through the fits in list and extract Survival Estimates information
   based on input parameters
       \enumerate{
            \item Retrieve survival estimate by calling the following function:\cr
                  summary(fit,tidy=TRUE,start=0,ci=`ci`,se=`se`,B=`B`,
                  na.action=`na.action`,cl=`cl`,type = "survival",t=`t`)
                  note: standard error and confidence intervals are generated
                   only if inputs are TRUE
                  note: set seed before calling summary method for reproducibility
                  with set.seed(`seed`)
            \item Retrieve hazard estimates by calling the following function:\cr
                  summary(fit,tidy=TRUE,start=0,ci=`ci`,se=`se`,B=`B`,
                  na.action=`na.action`,cl=`cl`,type = "hazard",t=`t`)
                  note: standard error and confidence intervals are generated
                   only if inputs are TRUE
                  note: set seed before calling summary method for reproducibility
                  with set.seed(`seed`)
            \item Combine survival and hazard estimate by left merging survival
                  with hazard by time point and add suffix _hazard to hazard
                   estimates column names
            \item Add column param_dist with fit names 
            \item Add column distr_desc with updated fit names from step 2
        }
  \item Format and append all data from all distributions into data frame
  \item Add surv_formula, preprocessed_data and t as attributes with same name
        surv_formula and preprocessed_data are retrieve form fits attributes
  \item Return list
  }
}
\if{html}{
The contents of this section are shown in PDF user manual only.
}
}

