% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_extract_from_fit.R
\name{extract_gof}
\alias{extract_gof}
\title{Extract Goodness of Fit}
\usage{
extract_gof(
  fits = list(),
  preprocessed_data = attributes(fits)$preprocessed_data,
  exp = TRUE,
  digits = 2,
  custom_model_label = TRUE
)
}
\arguments{
\item{fits}{{\code{list}}\cr flexsurvreg fit object}

\item{preprocessed_data}{{\code{data.frame}}\cr output of preprocess_data function}

\item{exp}{{\code{logical}}\cr Boolean to indicate if exponential transformation
should be used on knots location values}

\item{digits}{{\code{integer}}\cr Indicate the number of decimal places to use when
rounding knots location values see \code{\link[base:Round]{base::round()}}}

\item{custom_model_label}{{\code{logical}}\cr Boolean to indicate if spline knots
locations should be included in the model label}
}
\value{
List with variance covariance information for each element of input list
}
\description{
Extract Goodness of Fit information from fit objects.
}
\section{Specification}{

\if{latex}{
Extract Goodness of Fit information from fit objects. 

Logical Specifications:
  \enumerate{
  \item Bulletproof check fit object list using \code{\link{bullet_proof}}
  \item Update name of list elements using \code{\link{add_info_fit_desc}} with
  parameter exp and digits
  \item Loop through the fits in list and extract variance covariance 
  information based on class of object
        \enumerate{
        \item For flexsurvreg and flexsurvspline models, extract Goodness of Fit
              (AIC and BIC) from result of flexsurv::glance(fit)  
        \item Create tibble with 3 columns: distr_desc with name from step 2,
              AIC and BIC
        }
  \item Return list
  }
}
\if{html}{
The contents of this section are shown in PDF user manual only.
}
}

