% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_rtf_meta.R
\name{fct_rtf_meta}
\alias{fct_rtf_meta}
\title{Customize RTF formatting of the outputs}
\usage{
fct_rtf_meta(
  output_suffix = list("all"),
  rtf_title_pop = c("Subtitle 1", "(ITT Population)"),
  rtf_footnote_pop = c("Database Cutoff Date: XXX"),
  time_horizon = c("20 Years"),
  therapy_data_var = list(name = "TRT01P", name_label =
    "Planned Treatment for Period 01", levels = NULL, labels = NULL),
  therapy_arm = levels(mkanr::adtte[["TRT01P"]]),
  param = "Overall Survival",
  models_title = "Standard Parametric",
  output_suffix_update = NULL,
  rtf_param = NULL,
  modified_value = list(NULL)
)
}
\arguments{
\item{output_suffix}{List containing suffixes of the outputs as per documentation, e.g.: "e0stdflex0gof" for Goodness of Fit table. output_suffix="all" will output entire list of formats for all outputs}

\item{rtf_title_pop}{Table subtitle(s) for all rtf outputs}

\item{rtf_footnote_pop}{Table footnote(s) for all rtf outputs}

\item{time_horizon}{Time horizon length for extrapolation}

\item{therapy_data_var}{{\code{list}}\cr therapy variable factor metadata with elements:
\tabular{lll}{ \code{name} \tab {\code{character}} \tab  variable name to convert \cr
\code{levels} \tab {\code{character}} \tab preferred order of values taken by variable \cr
\code{labels} \tab {\code{character}} \tab preferred label of values taken by variable \cr
\code{name_label} \tab {\code{character}} \tab variable label \cr}}

\item{therapy_arm}{Therapy arm description}

\item{param}{Parameter label}

\item{models_title}{Model title}

\item{output_suffix_update}{List of outputs as output_suffix for whom the parameter values are to be updated. Please note that the values need to be valid output_suffices as per documentation}

\item{rtf_param}{List of RTF parameters for whom the values are to be updated. Please note that the parameters names need to be valid}

\item{modified_value}{List of modified values}
}
\value{
list of lists containing rtf metadata
}
\description{
This function contains standards metadata for RTF tables and figures.
}
\section{Specification}{

\if{latex}{ This function contains a list of formats for tables and figures 
in RTF format for {mkanr} outputs.
It allows the user to update the RTF formatting of the output such as titles, 
footnotes, column headers and columns widths as per project requirement.

Logical Specifications:
  \enumerate{
  \item Check input parameters
  \item Define a list called `rtf_meta` of RTF formats with default parameter values as per mock-up document 
  \item Modify the RTF parameter values in `rtf_meta` list based on input lists as per function definition if `output_suffix_update`, `rtf_param` and `modified_value` are not NULL
  \enumerate{ 
      \item Define a helper function `update_value` to update a parameter value in RTF formats list as per input lists `output_suffix_update`, `rtf_param` and `modified_value`.
      \item Update selected values of parameters and outputs in `rtf_meta` list by applying `pmap` to `update_value` function
  }
  \item Return list of rtf formatting for the outputs (`rtf_meta`)
  \enumerate{
      \item if `output_suffix == "all"`, return entire list of RTF formatting for all outputs (`rtf_meta`)
      \item else, return RTF formatting for selected output as per output_suffix (`rtf_meta$output_suffix`)  
  }
}
}
\if{html}{
The contents of this section are shown in PDF user manual only.
}
}

\examples{
\dontrun{
#Generate default metadata object
rtf_metadata<-fct_rtf_meta(output_suffix=list("all"),
rtf_title_pop=c("Subtitle 1", "(Population)"),
rtf_footnote_pop=c("Database Cutoff Date: XXX"),
param="Overall Survival",
therapy_data_var = list(name="TRT01P",
name_label="Planned Treatment for Period 01",levels=NULL,labels=NULL),
time_horizon=c("20 Years"),
models_title="Standard Parametric",
therapy_arm = c("B: Placebo", "A: Drug X")
                             )
                             
#Look up Goodness of Fit table metadata
rtf_metadata$e0stdflex0gof

#Customized Goodness of Fit table metadata
rtf_meta<-fct_rtf_meta(output_suffix=list("all"),
rtf_title_pop=c("Subtitle 1", "(Population)"),
rtf_footnote_pop=c("Database Cutoff Date: XXX"),
 param="Overall Survival",
 time_horizon=c("20 Years"),
 therapy_data_var = list(name="TRT01P",
 name_label="Planned Treatment for Period 01",levels=NULL,labels=NULL),
 models_title="Standard Parametric",
 therapy_arm = c("B: Placebo", "A: Drug X"),
 output_suffix_update = list("e0stdflex0gof", "e0stdflex0gof", "e0stdflex0gof"),
 rtf_param=list("rtf_title[1]","header_label", "csv_output"),
 modified_value=list("Goodness of Fit Statistics Resulting from Independent
 Standard Parametric Modeling for Overall Survival in Placebo Arm",
 "Survival Model|AIC|BIC|Mean AIC BIC|Rank AIC|Rank BIC|Mean Rank AIC BIC", "FALSE")
                             )    
 }
 
}
