% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_format.R
\name{format_est}
\alias{format_est}
\title{Format Survival Estimates and include Kaplan-Meier estimates}
\usage{
format_est(
  est,
  preprocessed_data = attributes(est)$preprocessed_data,
  preprocessed_data_km = attributes(est)$preprocessed_data,
  surv_formula = attributes(est)$surv_formula,
  type = "fleming",
  ...,
  rename_output = "test",
  outdata_path = tempdir()
)
}
\arguments{
\item{est}{{\code{list}}\cr named list with survival estimates, each element represent
a model  (result from extract_est function)}

\item{preprocessed_data}{{\code{data.frame}}\cr data used to produced flexsurv estimates}

\item{preprocessed_data_km}{{\code{data.frame}}\cr data used to produced Kaplan-Meier
estimates (see \code{\link[survival:survfit]{survival::survfit()}})}

\item{surv_formula}{{\code{formula}}\cr formula used to produced Kaplan-Meier
estimates (see \code{\link[survival:survfit]{survival::survfit()}})}

\item{type}{{\code{character}}\cr type used to produced Kaplan-Meier
estimates (see \code{\link[survival:survfit]{survival::survfit()}})}

\item{...}{other parameters to pass to survfit (see \code{\link[survival:survfit]{survival::survfit()}})}

\item{rename_output}{{\code{character}}\cr output name}

\item{outdata_path}{{\code{character}}\cr location to save output}
}
\value{
{\code{data.frame}}\cr formatted data with variables:
\tabular{lll}{ \code{x_primary} \tab {\code{double}} \tab time information \cr
\code{y_primary} \tab {\code{double}} \tab original survival estimates \cr
\code{param_dist} \tab {\code{character}} \tab parameter description \cr
\code{group_primary} \tab {\code{factor}} \tab  distribution description \cr
\code{covariate(s)} \tab {\code{factor}} \tab covariate(s) as in right hand side of model formula \cr}
}
\description{
Format Survival Estimates to:
\itemize{
\item combine all models
\item add Kaplan-Meier estimates using either object attributes or additional inputs
\item apply formula based on Kaplan-Meier estimates and flexsurv estimates
}
}
\section{Specification}{

\if{latex}{
Format Survival Estimates and include Kaplan-Meier estimates.

Format Survival Estimates to:
- combine all models
- add Kaplan-Meier estimates using either object attributes or additional inputs
- apply formula based on Kaplan-Meier estimates and flexsurv estimates

Logical Specifications:
  \enumerate{
  \item Check input parameters using \code{\link{bullet_proof}}
  \item Convert est input to data.frame using do.call(dplyr::bind_rows,est)
  \item Generate survfit object using survival::survfit(surv_formula,
        data = preprocessed_data_km, type = type, ...)
  \item Retrieve analysis variable name from surv_formula using surv_formula[[2]][[2]]
  \item Extract survival estimates from survfit object with times from 0 to 
        max value of preprocessed_data_km$analysisvariable using summary method
  \item Combine survfit estimates with flexsurv estimates
        \enumerate{
        \item Decompose survfit strata variable to identify covariates
        \item If at least one covariate found
              \enumerate{
              \item Format strata variable from survfit strata to match est 
              output
              \item Format survfit summary to match flexsurv summary using
                    param_dist="kaplan_meier"
                    distr_desc="Kaplan-Meier"
                    time=time
                    est=surv
                    keep strata
              }
        \item If no covariate found
              \enumerate{
              \item Format survfit summary to match flexsurv summary using
                    param_dist="kaplan_meier"
                    distr_desc="Kaplan-Meier"
                    time=time
                    est=surv
                    keep strata
              }
        \item Append survfit estimates with estimates from previous step
        }
  \item Convert distr_desc variable to factor
  \item Save RData data to outdata_path with rename_output name
  \item Return data.frame
  }
}
\if{html}{
The contents of this section are shown in PDF user manual only.
}
}

\examples{
\dontrun{
# regular analysis
tte <- preprocess_data(population_from = mkanr::adsl,
                observation_from = mkanr::adtte)
                
tte |>
   run_models_flexsurvreg() |>
   extract_est()|>
   format_est()
   
# piece-wise analysis
tte <- preprocess_data(population_from = mkanr::adsl,
                      observation_from = mkanr::adtte,
                      observation_where = AVAL2>10,
                      analysis_variable=list(name="AVAL2", derivation="AVAL/7-10",
                                             label='Analysis Variable 2'),
                      rm_obs=TRUE)

tte_km<- preprocess_data(population_from = mkanr::adsl,
                      observation_from = mkanr::adtte,
                      observation_where = ,
                      analysis_variable=list(name="AVAL2", derivation="AVAL/7",
                                             label='Analysis Variable 2'),
                      rm_obs=FALSE)

tte |>
 run_models_flexsurvreg() |>
 extract_est()|>
 format_est(preprocessed_data_km=tte_km)
                
}
}
