% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_format.R
\name{format_gof}
\alias{format_gof}
\title{Format Goodness of Fit}
\usage{
format_gof(
  gof,
  gof_arrange = NULL,
  rename_output = "test",
  outdata_path = tempdir()
)
}
\arguments{
\item{gof}{{\code{list}}\cr output of extract_gof function}

\item{gof_arrange}{{\code{character}}\cr variable used to order output. Possible
values are:
- NULL = no ordering
- 'Rank AIC'
- 'Rank BIC'
- 'Rank Mean AIC BIC'}

\item{rename_output}{{\code{character}}\cr output name}

\item{outdata_path}{{\code{character}}\cr location to save output}
}
\value{
{\code{data.frame}}\cr formatted data with variables:
\tabular{lll}{ \code{Survival Model} \tab {\code{character}} \tab Model distribution description \cr
\code{AIC} \tab {\code{double}} \tab Akaike Information Criterion  \cr
\code{BIC} \tab {\code{double}} \tab Bayesian Information Criterion \cr
\code{Mean AIC BIC} \tab {\code{double}} \tab Mean of AIC and BIC \cr
\code{Rank AIC} \tab {\code{double}} \tab Akaike Information Criterion ranking  \cr
\code{Rank BIC} \tab {\code{double}} \tab Bayesian Information Criterion ranking \cr
\code{Rank AIC BIC} \tab {\code{double}} \tab Mean AIC BIC ranking \cr}
}
\description{
Format Goodness of Fit to:
\itemize{
\item combine all models
\item add Mean AIC BIC, rank AIC, rank BIC and rank Mean AIC BIC
\item order table by one of the variable above
}
}
\section{Specification}{

\if{latex}{
Format Goodness of Fit 

Format Goodness of Fit to:
- combine all models
- add Mean AIC BIC, rank AIC, rank BIC and rank Mean AIC BIC
- order table by one of the variable above

Logical Specifications:
  \enumerate{
  \item Check input parameters using \code{\link{bullet_proof}}
  \item Convert extract_gof output to data.frame using 
  do.call(dplyr::bind_rows,gof)
  \item Derive additional variables
       \enumerate{
       \item Mean AIC BIC = (AIC + BIC)/2
       \item Rank AIC = base::rank(AIC)
       \item Rank BIC = base::rank(BIC)
       \item Rank Mean AIC BIC = base::rank(`Mean AIC BIC`)
       }
  \item Order table gof_arrange is a valid column name
  \item Save RData data to outdata_path with rename_output name
  \item Return data
  }
}
\if{html}{
The contents of this section are shown in PDF user manual only.
}
}

