% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_run_models.R
\name{h_main_flexsurvreg}
\alias{h_main_flexsurvreg}
\title{h_main_flexsurvreg main functionality}
\usage{
h_main_flexsurvreg(dist, surv_formula, preprocessed_data, na.action, ...)
}
\arguments{
\item{dist}{distribution parameter used in flexsurv::flexsurvreg()}

\item{surv_formula}{{\code{formula}}\cr Surv formula define with stats::formula and survival::Surv}

\item{preprocessed_data}{{\code{data.frame}}\cr Output of preprocess_data function}

\item{na.action}{a missing-data filter function, applied after any 'subset' argument has been used.
Default is na.omit}

\item{...}{other parameter to pass in flexsurv::flexsurvreg}
}
\description{
h_main_flexsurvreg main functionality
}
\keyword{internal}
