% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_run_models.R
\name{h_main_flexsurvspline}
\alias{h_main_flexsurvspline}
\title{h_main_flexsurvspline main functionality}
\usage{
h_main_flexsurvspline(
  scalexk,
  preprocessed_data,
  surv_formula,
  seed_value,
  method,
  na.action,
  ...
)
}
\arguments{
\item{scalexk}{list with scale and knot arguement used in flexsurv::flexsurvreg()}

\item{preprocessed_data}{{\code{data.frame}}\cr Output of preprocess_data function}

\item{surv_formula}{{\code{formula}}\cr Surv formula define with stats::formula and survival::Surv}

\item{seed_value}{Value for seed to be used in set.seed function}

\item{method}{Method to be used (refer to stats::optim Details section)}

\item{na.action}{a missing-data filter function, applied after any 'subset' argument has been used.
Default is na.exclude}

\item{...}{other parameters to use in flexsurv::flexsurvspline}
}
\description{
h_main_flexsurvspline main functionality
}
\keyword{internal}
