% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{mkanr-package}
\alias{mkanr}
\alias{mkanr-package}
\title{mkanr: Demo R package}
\description{
Demo Package to illustrate how to leverage on the layered design.
}
\author{
\strong{Maintainer}: Jean Muller \email{jean.muller@msd.com}

Authors:
\itemize{
  \item Ligia Adamska \email{ligia.adamska@msd.com}
}

Other contributors:
\itemize{
  \item BARDS, MSD [copyright holder]
}

}
\keyword{internal}
