% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot_smooth_line.R
\name{plot_smooth_line}
\alias{plot_smooth_line}
\title{Create line plot}
\usage{
plot_smooth_line(
  data,
  xvar = "x_primary",
  yvar = "y_primary",
  levelvar = "group_primary",
  facetvar = NULL,
  break_point = NULL,
  rename_output = "test",
  xlab = "X label",
  ylab = "Y label",
  fig_res = 300,
  fig_height = 4,
  fig_width = 6.6,
  fig_scale = 1.3,
  fig_line_size = 0.7,
  fig_grid_x = TRUE,
  fig_grid_y = TRUE,
  fig_scale_x_continuous = NULL,
  fig_scale_y_continuous = NULL,
  palette = c("#00857C", "#6ECEB2", "#BFED33", "#FFF063", "#0C2340", "#5450E4",
    "#688CE8", "#69B8F7", "#799A0E", "#BFED33", "#66203A", "#DB8DAA", "#F68D2E",
    "#FABB82", "#00857C", "#6ECEB2", "#BFED33", "#FFF063", "#0C2340", "#5450E4",
    "#688CE8", "#69B8F7", "#799A0E", "#BFED33"),
  legend_position = "bottom",
  facetvar_legend = ggplot2::labeller(.cols = base::identity),
  levelvar_legend = "Type",
  legend_textsize = 10,
  outgraph_path = tempdir(),
  outdata_path = tempdir()
)
}
\arguments{
\item{data}{Data for the line plot}

\item{xvar}{Independent variable}

\item{yvar}{Dependent variable}

\item{levelvar}{Category variable}

\item{facetvar}{Strata variable (for ggplot2::facet_grid)}

\item{break_point}{X-axis value to indicate break point value for piece-wise analysis}

\item{rename_output}{Name of output files}

\item{xlab}{X-axis label}

\item{ylab}{Y-axis label}

\item{fig_res}{Resolution (Pixels per inch) of the figure}

\item{fig_height}{Height (inch) of the figure}

\item{fig_width}{Width (inch) of the figure}

\item{fig_scale}{Scale factor of the figure}

\item{fig_line_size}{Line width in mm}

\item{fig_grid_x}{Logical. Show x axis grid lines (vertical)}

\item{fig_grid_y}{Logical. Show y axis grid lines (horizontal)}

\item{fig_scale_x_continuous}{Customize scale of x-axis (see ggplot2::scale_x_continuous())}

\item{fig_scale_y_continuous}{Customize scale of y-axis (see ggplot2::scale_y_continuous())}

\item{palette}{Palette to use for plotting}

\item{legend_position}{Position of the legend
Possible values are ("none", "left", "right", "bottom", "top")}

\item{facetvar_legend}{Title of the legend for facetvar (see ggplot2::labeller())}

\item{levelvar_legend}{Title of the legend for levelvar}

\item{legend_textsize}{Text size of the legend}

\item{outgraph_path}{Directory for the .png output}

\item{outdata_path}{Directory for the .RData output}
}
\value{
ggplot2 object
}
\description{
Creates a line plot and saves it to .RData and .png format.
}
\section{Specification}{

\if{latex}{ This function creates a line plot and saves it to .RData and .png format.

Logical Specifications:
 
  \enumerate{
  \item Check validity of input parameters
  \item Create the plot
  \item Save plot and associated data in outdata and .png in outgraph
  \item Return plot
  }

}
\if{html}{
The contents of this section are shown in PDF user manual only.
}
}

