% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_preprocess_data.R
\name{preprocess_data}
\alias{preprocess_data}
\title{Preprocess Data for Analysis}
\usage{
preprocess_data(
  population_from = adsl,
  population_where = TRUE,
  observation_from = adtte,
  observation_where = TRUE,
  analysis_variable = list(name = "AVAL2", derivation = "AVAL/7", label =
    "Analysis Variable 2"),
  analysis_variable_unit = list(name = "AVAL2U", derivation = "'WEEKS'", label =
    "Analysis Variable 2 Unit"),
  therapy_data_var = list(name = "TRT01P", name_label =
    "Planned Treatment for Period 01", levels = NULL, labels = NULL)
)
}
\arguments{
\item{population_from}{path to population SAS data or data.frame}

\item{population_where}{quosure or logical vector for where clause}

\item{observation_from}{path to observation SAS data or data.frame}

\item{observation_where}{quosure or logical vector for where clause}

\item{analysis_variable}{{\code{list}}\cr analysis variable derivation metadata with elements:
\tabular{lll}{ \code{name} \tab {\code{character}} \tab  variable name \cr
\code{derivation} \tab {\code{character}} \tab variable derivation \cr
\code{label} \tab {\code{character}} \tab variable label \cr}}

\item{analysis_variable_unit}{{\code{list}}\cr analysis variable unit derivation metadata with elements:
\tabular{lll}{ \code{name} \tab {\code{character}} \tab  variable name \cr
\code{derivation} \tab {\code{character}} \tab variable derivation \cr
\code{label} \tab {\code{character}} \tab variable label \cr}}

\item{therapy_data_var}{{\code{list}}\cr therapy variable factor metadata with elements:
\tabular{lll}{ \code{name} \tab {\code{character}} \tab  variable name to convert \cr
\code{levels} \tab {\code{character}} \tab preferred order of values taken by variable \cr
\code{labels} \tab {\code{character}} \tab preferred label of values taken by variable \cr
\code{name_label} \tab {\code{character}} \tab variable label \cr}
if levels and labels are set to NULL, sort(unique(name)) will be
used for both
see \code{\link[base:factor]{base::factor()}} for details}
}
\value{
A data.frame of class \code{`tibble`} containing pre processed data
ready to be used in modelling functions. The pre-processing includes merging
of population and observation data, filtering, derivation of analysis variables
and creation of factor for treatment variable.
}
\description{
Uses input data and parameters to generate pre-processed data.
}
\details{
For more information on quosures go to https://adv-r.hadley.nz/evaluation.html#quosures
}
\section{Specification}{

\if{latex}{ This function uses input data and parameters to generate pre-processed
data ready for analysis. Function assumes time variable is available in days 
in observation_from.

Logical Specifications:

  \enumerate{
  \item TryCatch to identify x_where condition type
  \item Preprocess quosures for x_where conditions
  \item Check input parameters using \code{\link{bullet_proof}}
  \item Get the data from population_from and filter by population_where.
     \enumerate{
        \item if population not a data.frame, read in population_from data with read_sas function
        \item filter data with population_where condition
        \item convert empty character values to missing (NA) values
        \item convert therapy variable to factor with information provided in
           therapy_data_var
        }
  \item Get the data from observation_from, derive analysis variable and filter by observation_where 
     \enumerate{
        \item if observation not a data.frame, read in observation_from data with read_sas function
        \item convert empty character values to missing (NA) values
        \item derive analysis variable and associated unit variable based on input
          analysis_variable and analysis_variable_unit (name and derivation elements)
        \item assign analysis variable and associated unit variable labels based on 
          input analysis_variable and analysis_variable_unit (label element)
       \item filter data with observation_where condition
        \item check if participants in population not present in observation 
        }
  \item Combine population and observation data
     \enumerate{
        \item left join population data with observation data and keep all variables
            except duplicates
        }
  \item Check result is a non-empty data.frame
  \item Return pre-processed data
  }

}
\if{html}{
The contents of this section are shown in PDF user manual only.
}
}

\examples{
\dontrun{
#example using path and quosures
data1<-preprocess_data(
population_from = mkanr::adsl,
population_where = quo(ITTFL=="Y" & TRT01P=="B: Placebo"),
observation_from = mkanr::adtte,
observation_where =  quo(PARAMCD=="OS") ,
analysis_variable = list(name="AVAL2", derivation="AVAL/7", label='Analysis Variable 2'),
analysis_variable_unit = list(name="AVAL2U", derivation="'WEEKS'", label='Analysis Variable 2 Unit')
)

}
}
