% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_rdata_to_rtf.R
\name{rdata_to_rtf}
\alias{rdata_to_rtf}
\title{Create RTF table from .RData file using r2rtf package}
\usage{
rdata_to_rtf(
  rename_output,
  rtf_title = c("Title 1", "(Population)"),
  rtf_footnote = c("{^a} Footnote 1", "{^b} Footnote 2"),
  rel_col_widths = c(2.5, rep(1, 3)),
  col_just = c("l", rep("c", 3)),
  col_format = rep(NA, 4),
  col_digits = rep(NA, 4),
  header_label = "",
  csv_output = TRUE,
  orientation = "portrait",
  outdata_path = path$outdata,
  outtable_path = path$outtable
)
}
\arguments{
\item{rename_output}{Name for output files}

\item{rtf_title}{Table title(s) for rtf output (Default c("Title 1","(Population)"))}

\item{rtf_footnote}{Table footnote(s) for rtf output (Default c("{^a} Footnote 1","{^b} Footnote 2"))}

\item{rel_col_widths}{Relative columns widths (Default c(2.5,rep(1,3)))}

\item{col_just}{Columns justification (Default c("l", rep("c", 3)))}

\item{col_format}{Number formatting as per format parameter of formatC function, use NA for string or no formatting (Default rep(NA,4))}

\item{col_digits}{The desired number of digits after the decimal point as per digits parameter of formatC function, use NA for no formatting (Default rep(NA,4))}

\item{header_label}{Columns custom label (Default "")}

\item{csv_output}{Boolean (TRUE/FALSE) to indicate if csv output should be produced (Default TRUE)}

\item{orientation}{Orientation in 'portrait' or 'landscape' (Default "portrait")}

\item{outdata_path}{Directory for the .RData table (Default path$outdata)}

\item{outtable_path}{Directory for the .rtf table (Default path$outtable)}
}
\value{
A text containing the path to the created RTF table.
}
\description{
Creates an RTF table with title and footer from an RData output using r2rtf package.
}
\section{Specification}{

\if{latex}{ This function creates an RTF plot with title and footer from an RData output using r2rtf package.

 Logical Specifications:

  \enumerate{
  \item Check input parameters
  \item If csv_output equals TRUE, use write.csv to generate un-formatted csv file in outdata folder
  \item Assemble the title, footnote and .RData output into an RTF table output in outtable folder
  }
  
}
\if{html}{
The contents of this section are shown in PDF user manual only.
}
}

