% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_run_models.R
\name{run_models_flexsurvreg}
\alias{run_models_flexsurvreg}
\title{Flexsurvreg Wrapper}
\usage{
run_models_flexsurvreg(
  preprocessed_data,
  surv_formula = stats::formula(survival::Surv(AVAL2, 1 - CNSR) ~ 1),
  dist_list = list(exponential = "exp", weibull = "weibull", gompertz = "gompertz", gamma
    = "gamma", lognormal = "lognormal", loglogistic = "llogis", gengamma = "gengamma"),
  na.action = na.omit,
  ...
)
}
\arguments{
\item{preprocessed_data}{{\code{data.frame}}\cr Output of preprocess_data function}

\item{surv_formula}{{\code{formula}}\cr Surv formula define with stats::formula and survival::Surv}

\item{dist_list}{{\code{list}}\cr List of distributions of parametric model
see flexsurv::flesurvreg for details}

\item{na.action}{a missing-data filter function, applied after any 'subset' argument has been used.
Default is na.omit}

\item{...}{other parameter to pass in flexsurv::flexsurvreg}
}
\value{
Named list of flexsurvreg objects (see \code{\link[flexsurv:flexsurvreg]{flexsurv::flexsurvreg()}} for details)
}
\description{
\code{run_models_flexsurvreg()} is a wrapper around \code{\link[flexsurv:flexsurvreg]{flexsurv::flexsurvreg()}}.
It loops through dist_list to run flexible parametric models for the same set of
input parameters (preprocessed_data, surv_formula, na.action, ...) and
different distributions (see dist parameter from [flexsurv::flexsurvreg())].

[flexsurv::flexsurvreg())]: R:flexsurv::flexsurvreg())
}
\section{Specification}{

\if{latex}{
`run_models_flexsurvreg()` is a wrapper around [flexsurv::flexsurvreg()].
It loops through dist_list to run flexible parametric models for the same set of
input parameters (preprocessed_data, surv_formula, na.action, ...) and
different distributions (see dist parameter from [flexsurv::flexsurvreg)].

 Logical Specifications:
  \enumerate{
  \item Check input parameters using \code{\link{bullet_proof}}
  \item Loop through the distribution in dist_list (dist) and run [flexsurv::flexsurvreg()]
    function on dist and other input parameters (surv_formula, na.action, ...).
        \enumerate{
        \item inform user when a model starts running
        \item wrap try catch around model run to retrieve errors and warnings if any
        \item if flexsurvsreg throws any error/warning, inform user and do not
            include model in output
        \item if flexsurvreg does not throw error/warning, return model fit
            object
       }
   \item Create list of fits with elements named as in dist_list input
   \item Add the following attributes to the list of fits to allow traceability
   \enumerate{
        \item preprocessed_data (=preprocessed_data)
        \item surv_formula (=surv_formula)
        \item dist_list (=dist_list)
        \item na.action (=na.action)
        \item model_type (="flexsurvreg")
        \item other (=rlang::exprs(...))
       }
   \item If at least one model ran without error/warnings return named list 
   of flexsurvreg fit objects, throw an error otherwise 
   }
}
\if{html}{
The contents of this section are shown in PDF user manual only.
}
}

