% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_run_models.R
\name{run_models_flexsurvspline}
\alias{run_models_flexsurvspline}
\title{Flexsurvspline Wrapper}
\usage{
run_models_flexsurvspline(
  preprocessed_data,
  surv_formula = stats::formula(survival::Surv(AVAL2, 1 - CNSR) ~ TRT01P),
  scalexk_list = list(hazard_1k = list(scale = "hazard", k = 1), hazard_2k = list(scale =
    "hazard", k = 2), hazard_3k = list(scale = "hazard", k = 3), odds_1k = list(scale =
    "odds", k = 1), odds_2k = list(scale = "odds", k = 2), odds_3k = list(scale = "odds",
    k = 3), normal_1k = list(scale = "normal", k = 1), normal_2k = list(scale = "normal",
    k = 2), normal_3k = list(scale = "normal", k = 3)),
  seed_value = 7902,
  na.action = na.omit,
  method = "Nelder-Mead",
  ...
)
}
\arguments{
\item{preprocessed_data}{{\code{data.frame}}\cr Output of preprocess_data function}

\item{surv_formula}{{\code{formula}}\cr Surv formula define with stats::formula and survival::Surv}

\item{scalexk_list}{{\code{list}}\cr Named list of lists of scale and k parameters
see flexsurv::flexsurvspline for details}

\item{seed_value}{Value for seed to be used in set.seed function}

\item{na.action}{a missing-data filter function, applied after any 'subset' argument has been used.
Default is na.exclude}

\item{method}{Method to be used (refer to stats::optim Details section)}

\item{...}{other parameters to use in flexsurv::flexsurvspline}
}
\value{
Named list of flexsurvpline fit objects
}
\description{
\code{run_models_flexsurvpline()} is a wrapper around \code{\link[flexsurv:flexsurvspline]{flexsurv::flexsurvspline()}}.
It loops through scalexk_list to run flexible spline models for the same set of
input parameters  (preprocessed_data, surv_formula, na.action, method, seed.value ...)
and different scale and knots (see scale, k and knots parameter from [flexsurv::flexsurvspline())].

[flexsurv::flexsurvspline())]: R:flexsurv::flexsurvspline())
}
\section{Specification}{

\if{latex}{
`run_models_flexsurvpline()` is a wrapper around [flexsurv::flexsurvspline()].
It loops through scalexk_list to run flexible spline models for the same set of
input parameters  (preprocessed_data, surv_formula, na.action, method, seed.value ...)
and different scale and knots (see scale, k and knots parameter from [flexsurv::flexsurvspline())].

 Logical Specifications:
  \enumerate{
  \item Check input parameters using \code{\link{bullet_proof}}
  \item Loop through scalexk_list the number of knots in spline (k) and scale (hazard,
      odds or probit) and run [flexsurv::flexsurvspline())] function on a set
      of input parameters (preprocessed_data, surv_formula, na.action, method, seed.value ...)
        \enumerate{
        \item wrap try catch around model to retrieve errors and warnings
        \item if flexsurvspline throws any error/warning, inform user and do not
            include model output
        \item if flexsurvspline does not throw error/warning, return model fit
            object
       }
   \item Create list of fits with elements named as in scalexk_list input
   \item Add the following attributes to the list of fits to allow traceability
   \enumerate{
        \item preprocessed_data (=preprocessed_data)
        \item surv_formula (=surv_formula)
        \item scalexk_list (=scalexk_list)
        \item na.action (=na.action)
        \item model_type (="flexsurvspline")
        \item method (=method)
        \item seed_value (=seed_value)
        \item other (=rlang::exprs(...))
       }
   \item If at least one model ran without error/warnings return named list 
   of flexsurvspline fit objects, throw an error otherwise 
   }
}
\if{html}{
The contents of this section are shown in PDF user manual only.
}
}

